DatabaseAdministration/requirements.txt
Flask
Flask-BasicAuth
mysql-connector